Fixes Brazilian phone format
This fix accounts for the relatively stable format
of area codes in Brazil since number of grantable
cellphones are usually increased by adding another
digit in the number format. The format is
described in Resolution 263 from Anatel. 

Also, fixes the extra digit rule, in which it's
possible to have any digit after the extra 9.